Simplified tests
