--- tests/config.json
@@ -1,12 +1,0 @@
-{
-  "root": "localhost:8080",
-  "chrome_driver_path": "./tests",
-  "conditional_question": "Do you feel unsafe right now?",
-  "validation_message": "Select items that you need help with",
-  "halt_dialog": "Emergency",
-  "title": "techKnow (Dev)",
-  "null_result" :{
-    "title": "Positive outcome (title)",
-    "content": "Positive outcome (content)"
-  }
-}
